+++ source/react/library/tabs/Tabs.js
@@ -2,19 +2,23 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { LEFT_KEY_CODE, RIGHT_KEY_CODE, UP_KEY_CODE } from '../../constants';
+import withId from '../../helpers/withId';
 
 import Tab from './Tab';
 import Panel from './Panel';
 
 const propTypes = {
-  /** For accesibility, a unique ID is required */
+  /**
+   * This prop is automatically passed from the withID HOC
+   * @ignore
+   */
   id: PropTypes.string.isRequired,
   /** Currently controls bg color of active tab & panel */
   type: PropTypes.oneOf(['primary', 'secondary']),
   /** Nested Tab.Tabs components */
   children: PropTypes.node,
   /** Optionally set active Tab with Tab ID */
-  activeTab: PropTypes.string,
+  initialTab: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   /** Optional onChange event handler. If onChange exists, Tabs are in controlled mode */
   onChange: PropTypes.func,
   /** Optional additional className */
@@ -26,40 +30,43 @@ const propTypes = {
 const defaultProps = {
   type: 'primary',
   children: null,
-  activeTab: null,
-  onChange: null,
+  initialTab: null,
+  onChange() {},
   className: '',
   style: {},
 };
 
+const collectChildProps = children =>
+  React.Children.toArray(children)
+    .filter(child => child && child.props)
+    .map((child, index) => ({
+      ...child.props,
+      index,
+    }));
+
+const getTabId = ({ id, index }) => id || index;
+
 class Tabs extends React.Component {
   constructor(props) {
     super(props);
-    // By default, the first tab is active
-    const defaultActive = React.Children.toArray(props.children).filter(
-      child => child && child.props,
-    )[0];
+
+    const childProps = collectChildProps(props.children);
 
     this.state = {
-      activeTab: props.activeTab || (defaultActive && defaultActive.props.id),
-      dirty: false,
+      activeTab:
+        props.initialTab || (childProps.length ? getTabId(childProps[0]) : 0),
     };
 
     this.onClick = this.onClick.bind(this);
-    this.onKeyDown = this.onKeyDown.bind(this);
   }
 
   onClick(activeTab) {
     const { onChange } = this.props;
-
-    if (onChange) {
-      onChange(activeTab);
-    } else {
-      this.setState({ activeTab, dirty: true });
-    }
+    this.setState({ activeTab });
+    onChange(activeTab);
   }
 
-  onKeyDown(event) {
+  onKeyDown(event, childProps) {
     const key = event.keyCode;
     const isSwitched = key === LEFT_KEY_CODE || key === RIGHT_KEY_CODE;
     const offset = -(UP_KEY_CODE - key);
@@ -67,22 +74,26 @@ class Tabs extends React.Component {
     if (isSwitched) {
       event.preventDefault();
 
-      this.switchTabOnArrowPress(offset);
+      this.switchTabOnArrowPress(offset, childProps);
     }
   }
 
-  switchTabOnArrowPress(offset) {
-    const { children, onChange } = this.props;
+  getActiveTab(childProps) {
     const { activeTab } = this.state;
-    let activeIndex = 0;
 
-    React.Children.toArray(children)
-      .filter(child => child && child.props)
-      .forEach((child, index) => {
-        const { id } = child.props;
+    const activeChild = childProps.find(props => props.active);
 
-        if (activeTab === id) activeIndex = index;
-      });
+    return (activeChild && getTabId(activeChild)) || activeTab;
+  }
+
+  switchTabOnArrowPress(offset, childProps) {
+    const { children, onChange } = this.props;
+
+    const activeTab = this.getActiveTab(childProps);
+
+    const activeIndex = childProps.findIndex(
+      child => getTabId(child) === activeTab,
+    );
 
     const nextIndex = activeIndex + offset;
     let adjustedNextIndex;
@@ -95,52 +106,47 @@ class Tabs extends React.Component {
       adjustedNextIndex = 0;
     }
 
-    const newActiveTab = React.Children.toArray(children).filter(
-      child => child && child.props,
-    )[adjustedNextIndex].props.id;
+    const newActiveTab = getTabId(childProps[adjustedNextIndex]);
 
-    if (onChange) {
-      onChange(newActiveTab);
-    } else {
-      this.setState({ activeTab: newActiveTab, dirty: true });
-    }
+    this.setState({ activeTab: newActiveTab });
+
+    onChange(newActiveTab);
   }
 
   render() {
     const { children, className, style, type, id: tabsId } = this.props;
-    const { activeTab, dirty } = this.state;
+
+    const childProps = collectChildProps(children);
+
+    const activeTab = this.getActiveTab(childProps);
 
     const tabs = [];
     const panels = [];
 
-    React.Children.toArray(children)
-      .filter(child => child && child.props)
-      .forEach(child => {
-        const { id } = child.props;
-        const active = activeTab === id;
+    childProps.forEach(props => {
+      const id = getTabId(props);
+      const active = activeTab === id;
 
-        const getKey = component => `${tabsId}-${component}-${id}`;
+      const panelProps = {
+        ...props,
+        id,
+        tabsId,
+        active,
+        key: `${tabsId}-panel-${id}`,
+      };
 
-        const panelProps = {
-          ...child.props,
-          id,
-          tabsId,
-          active,
-          key: getKey('panel'),
-        };
+      panels.push(<Panel {...panelProps} />);
 
-        panels.push(<Panel {...panelProps} />);
+      const tabProps = {
+        ...panelProps,
+        onClick: this.onClick,
+        onKeyDown: e => this.onKeyDown(e, childProps),
+        focused: active,
+        key: `${tabsId}-tab-${id}`,
+      };
 
-        const tabProps = {
-          ...panelProps,
-          onClick: this.onClick,
-          onKeyDown: this.onKeyDown,
-          focussed: dirty && active,
-          key: getKey('tab'),
-        };
-
-        tabs.push(<Tab {...tabProps} />);
-      });
+      tabs.push(<Tab {...tabProps} />);
+    });
 
     return (
       <div
@@ -161,4 +167,6 @@ Tabs.defaultProps = defaultProps;
 
 Tabs.Tab = Tab;
 
-export default Tabs;
+export { Tabs as UnwrappedTabs };
+
+export default withId(Tabs);